Adding route to homepage and adding recipe data to it. Also made some changes to server to get routes to work.
