--- controllers/shoppingCart.js
@@ -1,57 +1,0 @@
-const router = require("express").Router();
-const { Recipe, User } = require("../models");
-const withAuth = require("../utils/auth");
-
-/** Get home page */
-router.get("/", async (req, res) => {
-  try {
-    const recipeData = await Recipe.findAll({
-      // include: [
-      //   {
-      //     model: Recipe,
-      //     attributes: ["recipe_name"],
-      //   },
-      // ],
-    });
-    const recipes = recipeData.map((recipe) => recipe.get({ plain: true }));
-
-    res.render("shoppingCart", {
-      recipes,
-      logged_in: req.session.logged_in,
-    });
-  } catch (error) {
-    // console.log(error);
-    res.status(500).json(error);
-  }
-});
-
-// Use withAuth middleware to prevent access to route
-router.get("/profile", withAuth, async (req, res) => {
-  try {
-    // Find the logged in user based on the session ID
-    const userData = await User.findByPk(req.session.user_id, {
-      attributes: { exclude: ["password"] },
-      include: [{ model: Recipe }],
-    });
-
-    const user = userData.get({ plain: true });
-
-    res.render("profile", {
-      ...user,
-      logged_in: true,
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.get("/login", (req, res) => {
-  // If the user is already logged in, redirect the request to another route
-  if (req.session.logged_in) {
-    res.redirect("/profile");
-    return;
-  }
-
-  res.render("login");
-});
-module.exports = router;
