+++ seeds/recipeData.json
@@ -1,7 +1,8 @@
 [
     {
         "recipe_name": "Strawberry Banana Stuffed Pancakes",
-        "recipe_description": "Instructions: \n Mix together the Greek yogurt, 1/4 cup banana, egg whites, and vanilla extract into a small bowl. \n In a separate bowl combine flour, baking soda, 1/2 teaspoon sugar, and salt.\n Combine both bowls and whisk until evenly mixed.\nFold in 5 fresh, chopped strawberries.\n Place 1/4-cup batter onto a non-stick skillet over medium heat and cook until bubbles start appearing through the top of the pancake. Flip (about 2 minutes on each side).\n For the sauce: Add 1/4 cup mashed banana, 5 frozen strawberries, and 1 teaspoon of sugar in a sauce pan (if you use fresh strawberries, add 1 tablespoon of water to the sauce pan). Cook on low, stirring frequently until thickened (about 5-7 minutes.)\n Evenly divide sauce over top of pancakes",
+        "recipe_description": "A delicious treat for breakfast",
+        "recipe_instructions": "Instructions: \n Mix together the Greek yogurt, 1/4 cup banana, egg whites, and vanilla extract into a small bowl. \n In a separate bowl combine flour, baking soda, 1/2 teaspoon sugar, and salt.\n Combine both bowls and whisk until evenly mixed.\nFold in 5 fresh, chopped strawberries.\n Place 1/4-cup batter onto a non-stick skillet over medium heat and cook until bubbles start appearing through the top of the pancake. Flip (about 2 minutes on each side).\n For the sauce: Add 1/4 cup mashed banana, 5 frozen strawberries, and 1 teaspoon of sugar in a sauce pan (if you use fresh strawberries, add 1 tablespoon of water to the sauce pan). Cook on low, stirring frequently until thickened (about 5-7 minutes.)\n Evenly divide sauce over top of pancakes",
         "number_servings": 2,
         "unit_price": 15.99,
         "category_id": 2
