+++ views/homepage.hbs
@@ -1,0 +1,24 @@
+<div>
+    {{#each recipes as |recipe|}}
+  <div class="row">
+  
+      <div class="col-sm-6 col-sm-4">
+        <div class="thumbnail">
+          <!--<img class="imag-responsive" src="{{recipe.imagePath}}" alt="..." /> -->
+          <div class="caption">
+            <h3>{{recipe.recipe_name}}</h3>
+            <p class="description"> {{recipe.recipe_description}}  </p>
+            <div class="clearfix">
+              <div class="price pull-lest">${{recipe.unit_price}}</div>
+              <a
+                href="#"
+                class="btn btn-success pull-right"
+                role="button"
+              >Add to cart</a></div>
+          </div>
+        </div>
+      </div>
+
+  </div>
+{{/each}}
+</div>
