+++ views/partials/header.hbs
@@ -16,7 +16,7 @@
       <ul class="navbar-nav me-auto mb-2 mb-lg-0">
         <li class="nav-item">
           <a class="nav-link active" aria-current="page" href="#">Shopping Cart
-            <span class="badge">{{session.shoppingCart.total_amount}}</span>
+            <!-- <span class="badge">{{session.shoppingCart.total_amount}}</span> -->
           </a>
         </li>
